BUG FIX: UPDATE USER WAS SHARED SERVICE ADDED BOOLEAN CHECK imageDeleteRequired TO CHECK IF IMAGE IS REQUIRED TO DELETE

+++ apps/quick-learn-backend/src/routes/users/users.service.ts
@@ -240,7 +240,7 @@ export class UsersService extends PaginationService<UserEntity> {
     });
   }
 
-  async updateUser(uuid: UserEntity['uuid'], payload: Partial<UserEntity>) {
+  async updateUser(uuid: UserEntity['uuid'], payload: Partial<UserEntity> , imageDeleteRequired : boolean) {
     const user = await this.findOne({ uuid });
 
     if (!user) {
@@ -248,11 +248,12 @@ export class UsersService extends PaginationService<UserEntity> {
     }
 
     // ON PROFILE CHANGE VERIFY IF LOGO HAS CHANGED AND PERVIOUS IMAGE IS NOT EMPTY STRING
-    if (
+    if (imageDeleteRequired &&
       user.profile_image !== payload.profile_image &&
       user.profile_image !== '' &&
       user.profile_image !== null
     ) {
+      console.log("ENTERED HERE");
       // DELETE OLD IMAGE FROM S3 BUCKET
       await this.FileService.deleteFiles([user.profile_image]);
     }
